+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.3.0
+
+* Buschtelefon remote endpoints are now connected via the local endpoint's outbound port
+* `Chain` now queues up unconsolidated (future or invalid) blocks for future use.
+* `Runner` now loads main chain into Buschtelefon on start.
+* The miner thread now passes on thread execution after mining to make sure
+  that the produced block can be digested by the queue (otherwise we may mine blocks twice).
+* We can now inquire all remote neighbors manually via `Runner#inquire_all_neighbors`
+
 ## 0.2.0
 
 * Better threading